+++ src/core/jsx-dev-runtime.ts
@@ -1,8 +1,7 @@
-import { Attributes } from "./element.js";
-import { Fragment, jsx } from "./jsx-runtime.js";
+import { jsx } from "./jsx-runtime.js";
 
 export * from "./jsx-runtime.js";
 
-export function jsxDEV(tagName: string | typeof Fragment, props: Attributes & { children?: unknown }): unknown {
-	return jsx(tagName, props);
+export function jsxDEV(tagName: any, props: any, key: any): unknown {
+	return jsx(tagName, props, key);
 }