+++ examples/src/components/text-input.tsx
@@ -3,7 +3,7 @@ import { Signal, isPending } from "@mxjp/gluon";
 export function TextInput(props: { value: Signal<string> }) {
 	return <input
 		type="text"
-		value={props.value}
+		prop:value={props.value}
 		disabled={isPending}
 		$input={event => {
 			props.value.value = (event.target as HTMLInputElement).value;